refactor: Clean up Layout and extract logic
