+++ src/components/LiveCursors/Project.tsx
@@ -1,7 +1,6 @@
 import { useEffect } from "react";
 import { useOutletContext } from "react-router-dom";
-
-import type { ProjectInfo } from "../../Layout";
+import type { ProjectInfo } from "../commonUtils/types";
 import LiveCursors from "./LiveCursors";
 
 const Project = () => {
