+++ src/components/commonUtils/types.d.ts
@@ -1,0 +1,5 @@
+export type ProjectInfo = {
+  name: string;
+  repoNameAndPath: string;
+  topic: string;
+};
