+++ src/main.tsx
@@ -1,6 +1,5 @@
 import ReactDOM from "react-dom/client";
 import "./index.css";
-
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Layout from "./Layout";
 import LiveCursors from "./components/LiveCursors";
@@ -12,8 +11,8 @@ import UserClaims from "./components/UserClaims/UserClaims";
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <BrowserRouter>
     <Routes>
-      <Route path="/" element={<Home />} />
       <Route element={<Layout />}>
+        <Route path="/" element={<Home />} />
         <Route path="/live-cursors" element={<LiveCursors />} />
         <Route path="/avatar-stack" element={<AvatarStack />} />
         <Route path="/emoji-reactions" element={<EmojiReactions />} />
